+++ CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+### Added
+- Endpoint for multiple order retrieval (GET shopgate/v2/orders)
+- Endpoint for single order retrieval   (GET shopgate/v2/orders/:orderId)
+- Ability to retrieve order addresses/items/comments if Magento REST permissions are enabled
 
 ## [4.0.0] - 2019-02-25
 ### Changed